feat!: always write colon before scheme in SSURT
This commit changes the ssurt implementation to always write a colon before the scheme even when the port is empty.

The SSURT spec we follow is https://github.com/iipc/urlcanon/blob/master/ssurt.rst and clearly demonstrates in the examples that the colon should be present if there is no port.