refactor!: only accept single search key (url)
This commit changes the search API to only accept a single URL as the search key. Accepting two or more search keys was a workaround to enable searching for both HTTP and HTTPS at the same time.

This commit comes in preparation for a new database key format that enables searching for both HTTP and HTTPS without defering to searching in parallel with multpiple search keys.