feat: limit number of results for prefix searches
The maximum number of records that is returned when performing a prefix search
is now configurable (warcserver API).

If an exact match is not found when searching the index, pywb retries queries
using prefix matching without limit to the number of results. This often leads
to timeouts as the number of results can be large if the index is large.